Comments and refactoring.
